+++ legadilo/reading/tests/test_models/test_article.py
@@ -819,47 +819,85 @@ class TestArticleQuerySet:
         ]
 
     def test_for_cleanup(self, user, other_user):
-        ArticleFactory(title="Unread not linked to a feed (to keep)", user=user, read_at=None)
         ArticleFactory(
-            title="Read not linked to a feed (to keep)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Unread not linked to a feed (to keep)",
+            user=user,
+            read_at=None,
+            main_source_type=constants.ArticleSourceType.MANUAL,
+        )
+        ArticleFactory(
+            title="Read not linked to a feed (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.MANUAL,
         )
         feed_forever_retention = FeedFactory(user=user, article_retention_time=0)
         feed_one_day_retention = FeedFactory(user=user, article_retention_time=1)
         feed_seven_day_retention = FeedFactory(user=user, article_retention_time=7)
 
         unread_1_day_retention_to_keep = ArticleFactory(
-            title="Unread, linked to 1 day retention (to keep)", user=user
+            title="Unread, linked to 1 day retention (to keep)",
+            user=user,
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_one_day_retention.articles.add(unread_1_day_retention_to_keep)
 
         read_1_day_retention_to_cleanup = ArticleFactory(
-            title="Read, 1 day retention (to cleanup)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Read, 1 day retention (to cleanup)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_one_day_retention.articles.add(read_1_day_retention_to_cleanup)
 
+        read_1_day_retention_to_cleanup_but_was_manually_added_too = ArticleFactory(
+            title="Read, 1 day retention but was manually added too (to keep)!",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.MANUAL,
+        )
+        feed_one_day_retention.articles.add(
+            read_1_day_retention_to_cleanup_but_was_manually_added_too
+        )
+
         read_1_day_retention_to_keep = ArticleFactory(
-            title="Read, 1 day retention (to keep)", user=user, read_at=utcdt(2024, 6, 6, hour=12)
+            title="Read, 1 day retention (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 6, hour=12),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_one_day_retention.articles.add(read_1_day_retention_to_keep)
 
         read_seven_day_retention_to_keep = ArticleFactory(
-            title="Read, 7 days retention (to keep)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Read, 7 days retention (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_seven_day_retention.articles.add(read_seven_day_retention_to_keep)
 
         read_keep_forever_retention_to_keep = ArticleFactory(
-            title="Read keep forever (to keep)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Read keep forever (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_forever_retention.articles.add(read_keep_forever_retention_to_keep)
 
         read_keep_forever_and_one_day_retention_to_keep = ArticleFactory(
-            title="Read keep forever and 1 day (to keep)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Read keep forever and 1 day (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_forever_retention.articles.add(read_keep_forever_and_one_day_retention_to_keep)
         feed_one_day_retention.articles.add(read_keep_forever_and_one_day_retention_to_keep)
 
         read_keep_one_and_seven_days_retention_to_keep = ArticleFactory(
-            title="Read keep 1 and 7 days (to keep)", user=user, read_at=utcdt(2024, 6, 1)
+            title="Read keep 1 and 7 days (to keep)",
+            user=user,
+            read_at=utcdt(2024, 6, 1),
+            main_source_type=constants.ArticleSourceType.FEED,
         )
         feed_forever_retention.articles.add(read_keep_one_and_seven_days_retention_to_keep)
         feed_one_day_retention.articles.add(read_keep_one_and_seven_days_retention_to_keep)
@@ -1530,15 +1568,3 @@ class TestArticleModel:
         assert was_updated
         for attr, value in expected_data.items():
             assert getattr(article, attr) == value
-
-    @pytest.mark.parametrize(
-        ("source_type", "is_from_feed"),
-        [
-            pytest.param(constants.ArticleSourceType.FEED, True, id="feed"),
-            pytest.param(constants.ArticleSourceType.MANUAL, False, id="manual"),
-        ],
-    )
-    def test_is_from_feed(self, source_type, is_from_feed):
-        article = ArticleFactory.build(main_source_type=source_type)
-
-        assert article.is_from_feed == is_from_feed