+++ legadilo/reading/views/__init__.py
@@ -15,7 +15,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from .article_actions_views import (
-    delete_article_view,
     update_article_view,
 )
 from .article_details_views import article_details_view
@@ -35,7 +34,6 @@ from .search_views import search_view
 __all__ = [
     "add_article_view",
     "article_details_view",
-    "delete_article_view",
     "external_tag_with_articles_view",
     "reading_list_admin_view",
     "reading_list_create_view",