+++ legadilo/users/admin.py
@@ -87,5 +87,6 @@ class UserAdmin(auth_admin.UserAdmin):
 @admin.register(Notification)
 class NotificationAdmin(admin.ModelAdmin):
     readonly_fields = ("created_at", "updated_at")
+    autocomplete_fields = ("user",)
     list_display = ("title", "created_at", "user", "is_read")
     list_filter = ("is_read",)